requirements.txt
fastapi
uvicorn[standard]
httpx
gunicorn
python-dotenv
jinja2
